Put refiners in their own namespace
